Changed integer from a long to an int.

+++ include/value.h
@@ -6,7 +6,7 @@
 
 union valUnion {
     const char* string;
-    long integer;
+    int integer;
     float decimal;
     list array;
 };
